[codex:orchestration] extract unified-result kernel internals
