+++ sentientos/orchestration_spine/kernel/__init__.py
@@ -1,0 +1,5 @@
+"""Authority-kernel internal modules for orchestration spine decomposition."""
+
+from .unified_results import resolve_unified_orchestration_result_kernel
+
+__all__ = ["resolve_unified_orchestration_result_kernel"]
